requirements.txt
pytorch-lightning==1.9.0
PyYAML==6.0
tensorboard>=2.2.0
torchsummary==1.5.1